Refactor urlrequest
